socket_testing/python_example/requirements.txt
python-socketio
requests